Revert "APK"
This reverts commit 559957f8987c0a1f38647f7c36cb605faf35cc99

+++ app/src/main/java/cu/uci/coj/Problems/ProblemItem.java
@@ -58,7 +58,7 @@ public class ProblemItem implements Serializable{
         submit = JSONObject.getInt("sub");
         accept = JSONObject.getInt("ac");
         accept_percent = JSONObject.getInt("acporciento");
-        score = JSONObject.getDouble("score");
+        score = JSONObject.getInt("score");
         images = new ArrayList<>();
     }
